Contest location
The CDS & presentation clients support the location of the contest, identically to how organization locations are specified. This enables a World Map presentation where balloons fly from the home location of every team to the contest location for judgement.

IMHO this is generic enough - and optional! - that it's worth putting in the spec in case CCSs want to use this to drive the presentation directly, or have other uses for contest location.