Update notification format & feed/webhook
As discussed on the call today, the event feed was receiving a redundant contest_id and the webhook was inconsistent in whether the callback would have one or more notification objects (and what the format was if more than one). This contains a few changes:
- Revert notification `endpoint` to `type`. This is one less change from the current spec, focuses on the contest object vs the REST API, and avoids some minor terminology confusion.
- Remove `contest_id` attribute from notification format. For event-feed this is already implied (you can only add a webhook to one contest), and for webhook each callback should have an array of notifications for only one contest per callback.
- Change webhook callback to an object that contains contest_id and an array of notification objects for that contest.
- Updated and added examples.

This change keeps the nice property that it would be trivial to convert between webhook and event feed or vice-versa, e.g. if you save every webhook notification array into a file called <contest_id>/event-feed.json as NDJSON, you essentially have a feed.