Capabilities and Account endpoint
I looked at several REST APIs, and the pattern is often /users (for all users), /users/x for a specific user, and /user to get the current user. In more than one case /users/x and /user had different properties.

Following this same pattern, I'm proposing /account to get the currently logged in account. The properties are identical to /accounts where they overlap (id, username, team/people id, but it does not include all of the same properties (e.g. password, IP) and has a couple new ones: capabilities, and a list of types that tells you which types and properties for each you may see. The types list is shared with the notification format.

I've added a section to describe what capabilities are, but not attempt to define any of them. I'd like to focus on the accounts endpoint and capabilities description first, then use separate PR(s) to define the capabilities.