Inherently nullable properties
Changing all "inherently nullable" properties to actually allow null. These are all cases where we had a 'double no' for required and nullable before, or the property is only required in some cases. In all of them the client would already need to accept missing or null values (which we say are the same) or the server might have to do some odd things to support (e.g. have /access check if a contest has a formal_name to decide whether or not to include that property). Allowing this is easier for both parties.

Also fixed formatting of contest state table.