Remove unused account_ids
Remove account_id from submissions and clarifications before we declare this version of the spec, and do a few really minor fixes. The account was added with the expectation that we would use it at the last finals but plans changed; removing since it hasn't been used and better to add later when we can get feedback.