Multiple systems as subfolders
As discussed on the CLICS call I rewrote the section on multiple systems to change from using a naming postfix to separate subfolders. I moved it to its own section so that it could be expanded into multiple (short) paragraphs, and I tried to explain the reasoning a bit more.