Penalty time in RELTIME
Contest penalty_time is the only time that's in integer minutes, and one of only 3 places where we even say minutes in the spec. As per issue #150, it also blocks non-integer penalty times.

This changes it to use RELTIME, which improves both issues. It does not attempt to address anything else in issue 150, namely no attempt to add additional scoring algorithms.