Fix Issue #164 - Access Example uses undocumented capability
One of the examples for the Access endpoint showed a capability of "patch_time".  This capability does not appear in the Capabilities list that says "All capabilities are listed in the table below".  Clearly, "patch_time" is not in the list.  It makes sense to change the example to use a documented property and "contest_start" makes the most sense here.