Clarify maximum score for problems
This PR was started because we noticed that the problem format allows both
-infinity and +infinity as maximum scores, but the Contest API doesn't allow
either. After talking to Fredrik and Nicky we decided not to 'fix' this:
- -infinity is being removed from the next version of the problem format.
- +infinity is rarely used, and there's an easy workaround (just pick an
  arbitrarily high value) so we'll leave it.

While discussing this we found that the max_score was poorly defined as
'this is the max, but maybe not' and decided to simplify and make it more
strict. I went with just removing the weak phrase: this IS the maximum
score possible.

I also felt 'indicate' could be improved (went with 'determine') and
'in scoring contests' was redundant (removed it).

As with anything in the spec, we'd expect servers to enforce the max_score,
but encourage clients to be liberal in what they accept. e.g. currently
the resolver doesn't work if you score the maximum and then have another
submission, but we'll fix this.