Simplified judgement types
Adds an optional simplified judgement type id to judgement types. These would be
specified in cases where (e.g.) a team should not see the specific judgement that
another team received.

The length of the property is unfortunate, but follows the naming pattern elsewhere.

** This is the first half of this change - judgements may need to be modified as
well. We will not release a new version of the spec until we complete the second
change or back this one out.