Optional team_id for submissions
Some CCSs allow judges or other non-team accounts to make submissions, e.g. to test
the judging system or problem solutions prior to the start of a contest. This allows
team_id to be null for these situations, and adds an optional account_id to indicate
which account was used to make a submission.

First part of #230.